refactor: declared the type in a same file | SCRUM-313
